pyproject.toml
[tool.poetry]
name = "sherpa-tax-rule-studio"
version = "0.1.0"
description = "Tax law specification engine for authoring structured, machine-readable rule packages"
authors = ["Ken"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
django = "^5.2"
djangorestframework = "^3.15"
dj-database-url = "^2.3"
psycopg = {version = "^3.2", extras = ["binary"]}
python-dotenv = "^1.1"
gunicorn = "^23.0"
whitenoise = "^6.9"
django-cors-headers = "^4.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "server.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
